// src/app/core/api/models/app-schemas-conectividade-analise-temporal-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AnaliseTemporalData } from '../models/analise-temporal-data';

/**
 * Resposta da análise temporal por tipo de localização e dependência.
 */
export interface AppSchemasConectividadeAnaliseTemporalResponse {
  data: AnaliseTemporalData;

  /**
   * Identificador semântico do recurso.
   */
  descricao: string;
}