// src/app/core/api/models/app-schemas-conectividade-mapa-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AppSchemasConectividadeMapaData } from '../models/app-schemas-conectividade-mapa-data';

/**
 * Resposta do mapa: descrição + lista de pontos georreferenciados.
 */
export interface AppSchemasConectividadeMapaResponse {
  data: AppSchemasConectividadeMapaData;

  /**
   * Identificador semântico do mapa.
   */
  descricao: string;
}